requirements.txt
# Core PyTorch ecosystem
torch==2.0.1
torchvision==0.15.2

# Core ML/AI libraries
# Note: numpy<2.0.0 to avoid compatibility issues with detectron2/densepose
# DensePose may try to install numpy>=2.0, so we pin it explicitly
numpy>=1.24.0,<2.0.0
# Pin scipy to avoid numpy ufunc compatibility issues (scipy 1.15+ has issues with numpy 1.26)
scipy>=1.10.0,<1.12.0
scikit-image>=0.20.0
# Use opencv-python (not headless) and pin to version compatible with numpy<2.0
opencv-python>=4.8.0,<4.12.0
pillow>=10.0.0
diffusers==0.20.2
transformers>=4.30.0
xformers==0.0.20
accelerate>=0.20.0
matplotlib>=3.7.0
tqdm>=4.65.0
gradio==3.44.4
requests>=2.31.0
einops>=0.6.0
onnxruntime>=1.15.0

av>=10.0.0
cloudpickle>=2.2.0
fvcore>=0.1.5
# omegaconf version determined by hydra-core (>=2.2,<2.4)
hydra-core>=1.1.0,<1.4.0
pycocotools>=2.0.6

# Additional dependencies for detectron2 and densepose
setuptools>=65.0.0
wheel>=0.40.0
pyyaml>=6.0
yacs>=0.1.8
tabulate>=0.9.0
termcolor>=2.3.0
iopath>=0.1.9
portalocker>=2.7.0
pydot>=1.4.2

# Note: detectron2 and densepose will be installed from local source
# They are located in preprocess/detectron2/
basicsr